Rename mds-test-data/ladot-service-areas to mds-test-data/test-areas
