+++ packages/mds-test-data/test-areas/restricted-areas.js
